store one way salted hashed password

+++ custom/icds/tests/models/test_ccz_hosting_link.py
@@ -5,8 +5,8 @@ import six
 from django.test import TestCase
 from django.core.exceptions import ValidationError
 
-from corehq.motech.utils import b64_aes_encrypt
 from custom.icds.models import CCZHostingLink
+from custom.nic_compliance.utils import verify_password
 
 
 class TestCCZHostingLink(TestCase):
@@ -14,21 +14,27 @@ class TestCCZHostingLink(TestCase):
 
     def setUp(self):
         super(TestCCZHostingLink, self).setUp()
-        self.link = CCZHostingLink(username="user", domain="test", identifier="abcdef")
-        self.link.password = b64_aes_encrypt(self.raw_password)
+        self.link = CCZHostingLink(username="user", domain="test", identifier="abcdef",
+                                   password=self.raw_password)
+
+    def tearDown(self):
+        if self.link.pk:
+            self.link.delete()
 
     def test_valid_link(self):
         self.link.full_clean()
         self.link.save()
-        self.link.delete()
 
     def test_identifier_validation(self):
         self.link.identifier = "123-abd"
         with self.assertRaisesMessage(ValidationError, "must be lowercase alphanumeric"):
             self.link.full_clean()
 
-    def test_get_password(self):
-        self.assertEqual(self.link.get_password, self.raw_password)
+    def test_encrypted_password(self):
+        self.link.save()
+        self.assertNotEqual(self.link.password, self.raw_password)
+        self.assertTrue(verify_password(self.raw_password, self.link.password),
+                        "encrypted password does not match")
 
     def test_representation(self):
         self.assertEqual(six.text_type(self.link), self.link.identifier)
